code/requirements.txt
# Sistema de Detecção de Estado Operacional para Motobomba Industrial
# Dependências Python para o projeto TCC - Engenharia de Sistemas UFMG

# Core Machine Learning & Deep Learning
tensorflow==2.19.0
keras==3.9.2
scikit-learn==1.7.1

# Data Processing & Analysis
pandas==2.3.1
numpy==2.1.1
scipy==1.15.2

# Database & Time Series
influxdb==5.3.2

# Visualization
matplotlib==3.10.1
seaborn==0.13.2

# Model Persistence & Serialization
joblib==1.5.2
h5py==3.13.0

# Utilities & System
python-dateutil==2.8.2
pytz==2023.3

# GUI (Interface Gráfica)
customtkinter>=5.2.0
tkcalendar>=1.6.0
Pillow>=10.0.0

# Standard Library Extensions
typing-extensions>=4.0.0

# Optional: Jupyter for development and interactive analysis
# jupyter==1.0.0
# notebook==7.4.7
# ipykernel==7.1.0
# jupyterlab==4.4.10

# Optional: Development tools and testing
# pytest==8.4.2
# black==23.11.0
# flake8==6.1.0
# mypy==1.18.2
# ruff==0.14.4

# Optional: Additional utilities for advanced analysis
plotly==6.4.0
# tqdm>=4.0.0
